experiments/pyproject.toml
[project]
name = "dressme-experiments"
version = "0.1.0"
description = ""
requires-python = ">=3.13,<3.14"
dependencies = [
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "nbstripout>=0.9.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "replicate>=1.0.7",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
]
